server/src/modules/users/services/UpdateUserProfileService.spec.ts
import FakeUsersRepository from '@modules/users/repositories/fakes/FakeUsersRepository';
import AppError from '@shared/errors/AppError';
import FakeHashProvider from '../providers/HashProvider/fakes/FakeHashProvider';
import UpdateUserProfileService from './UpdateUserProfileService';

let fakeUsersRepository: FakeUsersRepository;
let fakeHashProvider: FakeHashProvider;
let updateProfile: UpdateUserProfileService;

describe('UpdateUserAvatar', () => {
  beforeEach(() => {
    fakeUsersRepository = new FakeUsersRepository();
    fakeHashProvider = new FakeHashProvider();

    updateProfile = new UpdateUserProfileService(
      fakeUsersRepository,
      fakeHashProvider,
    );
  });

  it('should be able to update the profile', async () => {
    const user = await fakeUsersRepository.create({
      username: 'johndoe',
      name: 'John Doe',
      email: 'john@gmail.com',
      password: '123456',
    });

    const updateUser = await updateProfile.execute({
      user_id: user.id,
      username: 'johndoe',
      name: 'John Tre',
      email: 'johntre@example.com',
    });

    expect(updateUser.name).toBe('John Tre');
    expect(updateUser.email).toBe('johntre@example.com');
  });

  it('should not be able to update the profile of a non existing user', async () => {
    expect(
      updateProfile.execute({
        user_id: 'non-existing-user-id',
        username: 'johndoe',
        name: 'name',
        email: 'email',
      }),
    ).rejects.toBeInstanceOf(AppError);
  });

  it('should not be able to update email to existing one', async () => {
    await fakeUsersRepository.create({
      username: 'johndoe',
      name: 'John Doe',
      email: 'john@gmail.com',
      password: '123456',
    });
    const user = await fakeUsersRepository.create({
      username: 'johndoe',
      name: 'John Doe',
      email: 'another@gmail.com',
      password: '123456',
    });

    await expect(
      updateProfile.execute({
        user_id: user.id,
        username: 'johndoe',
        name: 'John Doe',
        email: 'john@gmail.com',
      }),
    ).rejects.toBeInstanceOf(AppError);
  });

  it('should not be able to update username to existing one', async () => {
    await fakeUsersRepository.create({
      username: 'johndoe',
      name: 'John Doe',
      email: 'john@gmail.com',
      password: '123456',
    });
    const user = await fakeUsersRepository.create({
      username: 'johndoepast',
      name: 'John Doe',
      email: 'anotherjohn@gmail.com',
      password: '123456',
    });

    await expect(
      updateProfile.execute({
        user_id: user.id,
        username: 'johndoe',
        name: 'John Doe',
        email: 'anotherjohn@gmail.com',
      }),
    ).rejects.toBeInstanceOf(AppError);
  });

  it('should be able to update the password', async () => {
    const user = await fakeUsersRepository.create({
      username: 'johndoe',
      name: 'John Doe',
      email: 'john@gmail.com',
      password: '123456',
    });

    const updateUser = await updateProfile.execute({
      user_id: user.id,
      username: 'johndoe',
      name: 'John Tre',
      email: 'johntre@example.com',
      password: '123123',
      old_password: '123456',
    });

    expect(updateUser.password).toBe('123123');
  });

  it('should not be able to update the password without the old_password', async () => {
    const user = await fakeUsersRepository.create({
      username: 'johndoe',
      name: 'John Doe',
      email: 'john@gmail.com',
      password: '123456',
    });

    expect(
      updateProfile.execute({
        user_id: user.id,
        username: 'johndoe',
        name: 'John Tre',
        email: 'johntre@example.com',
        password: '123123',
      }),
    ).rejects.toBeInstanceOf(AppError);
  });

  it('should not be able to update the password with wrong old_password', async () => {
    const user = await fakeUsersRepository.create({
      username: 'johndoe',
      name: 'John Doe',
      email: 'john@gmail.com',
      password: '123456',
    });

    await expect(
      updateProfile.execute({
        user_id: user.id,
        username: 'johndoe',
        name: 'John Tre',
        email: 'johntre@example.com',
        password: '123123',
        old_password: '123123',
      }),
    ).rejects.toBeInstanceOf(AppError);
  });
});
